# docs/swagger.yaml
definitions:
...
  dto.LeaderboardResponse:
    properties:
      current_user:
        $ref: '#/definitions/dto.LeaderboardUserResponse'
      period:
        type: string
      top_users:
        items:
          $ref: '#/definitions/dto.LeaderboardUserResponse'
        type: array
    type: object
  dto.LeaderboardUserResponse:
    properties:
      level:
        type: integer
      rank:
        type: integer
      spirit_stage:
        type: integer
      spirit_type:
        type: string
      user_id:
        type: string
      username:
        type: string
      xp:
        type: integer
    type: object
...
  dto.LessonMediaResponse:
    properties:
      lesson_id:
        type: string
      media:
        additionalProperties:
          $ref: '#/definitions/dto.MediaAssetResponse'
        description: 'key: video, subtitle, thumbnail'
        type: object
    type: object
...
  dto.MediaAssetResponse:
    properties:
      duration:
        description: seconds
        type: integer
      file_size:
        type: integer
      height:
        type: integer
      id:
        type: string
      url:
        type: string
      width:
        type: integer
    type: object
  dto.MediaUploadResponse:
    properties:
      file_name:
        type: string
      file_size:
        type: integer
      file_type:
        type: string
      id:
        type: string
      url:
        type: string
    type: object
...
  shared.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/admin/lessons/{lessonId}/animation:
    post:
      consumes:
      - multipart/form-data
      description: Upload animation video file - Step 3 of production workflow (Admin
        only)
      parameters:
      - default: Bearer <admin_token>
        description: Admin Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Lesson ID
        in: path
        name: lessonId
        required: true
        type: string
      - description: Animation file (MP4, MOV, WEBM)
        in: formData
        name: animation
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.MediaUploadResponse'
              type: object
      security:
      - Bearer: []
      summary: Upload Lesson Animation (Admin)
      tags:
      - admin
      - production
  /api/v1/admin/lessons/{lessonId}/audio:
    post:
      consumes:
      - multipart/form-data
      description: Upload voice-over audio file - Step 2 of production workflow (Admin
        only)
      parameters:
      - default: Bearer <admin_token>
        description: Admin Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Lesson ID
        in: path
        name: lessonId
        required: true
        type: string
      - description: Audio file (MP3, WAV, AAC)
        in: formData
        name: audio
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.MediaUploadResponse'
              type: object
      security:
      - Bearer: []
      summary: Upload Lesson Audio (Admin)
      tags:
      - admin
      - production
  /api/v1/admin/lessons/{lessonId}/media:
    get:
      description: Get all media assets for a lesson (Admin only)
      parameters:
      - default: Bearer <admin_token>
        description: Admin Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Lesson ID
        in: path
        name: lessonId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LessonMediaResponse'
              type: object
      security:
      - Bearer: []
      summary: Get Lesson Media (Admin)
      tags:
      - admin
...
  /api/v1/admin/lessons/{lessonId}/subtitle:
    post:
      consumes:
      - multipart/form-data
      description: Upload subtitle file for lesson video (Admin only)
      parameters:
      - default: Bearer <admin_token>
        description: Admin Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Lesson ID
        in: path
        name: lessonId
        required: true
        type: string
      - description: Subtitle file (VTT, SRT)
        in: formData
        name: subtitle
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.MediaUploadResponse'
              type: object
      security:
      - Bearer: []
      summary: Upload Lesson Subtitle (Admin)
      tags:
      - admin
  /api/v1/admin/lessons/{lessonId}/thumbnail:
    post:
      consumes:
      - multipart/form-data
      description: Upload thumbnail image for lesson (Admin only)
      parameters:
      - default: Bearer <admin_token>
        description: Admin Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Lesson ID
        in: path
        name: lessonId
        required: true
        type: string
      - description: Thumbnail file (JPG, PNG, WEBP)
        in: formData
        name: thumbnail
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.MediaUploadResponse'
              type: object
      security:
      - Bearer: []
      summary: Upload Lesson Thumbnail (Admin)
      tags:
      - admin
...
  /api/v1/admin/media/assets/{assetId}:
    delete:
      description: Delete a media asset and its physical file (Admin only)
      parameters:
      - default: Bearer <admin_token>
        description: Admin Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Media Asset ID
        in: path
        name: assetId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - Bearer: []
      summary: Delete Media Asset (Admin)
      tags:
      - admin
  /api/v1/admin/media/statistics:
    get:
      description: Get statistics about media assets and usage (Admin only)
      parameters:
      - default: Bearer <admin_token>
        description: Admin Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      security:
      - Bearer: []
      summary: Get Media Statistics (Admin)
      tags:
      - admin
...
  /api/v1/leaderboard/all-time:
    get:
      consumes:
      - application/json
      description: Get all-time leaderboard rankings
      parameters:
      - description: Limit results (default 50)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LeaderboardResponse'
              type: object
      summary: Get All Time Leaderboard
      tags:
      - leaderboard
  /api/v1/leaderboard/monthly:
    get:
      consumes:
      - application/json
      description: Get monthly leaderboard rankings
      parameters:
      - description: Limit results (default 50)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LeaderboardResponse'
              type: object
      summary: Get Monthly Leaderboard
      tags:
      - leaderboard
  /api/v1/leaderboard/weekly:
    get:
      consumes:
      - application/json
      description: Get weekly leaderboard rankings
      parameters:
      - description: Limit results (default 50)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LeaderboardResponse'
              type: object
      summary: Get Weekly Leaderboard
      tags:
      - leaderboard
...
swagger: "2.0"